3722 -- Issue iwth interactive status bar
Started working on this and add the gotoSection which is attached to each link in the form-steps.  It is now very easy to go to the correct position.
However the cell phone is having a problem with the navigation.  I think the problem stem from the touchmove event since It is  using formScrollSpy.
FormScrollSpy is selecting the a link from the second set of navigation buttons in the form-steps, I notices that the last navigation links will
become active.   This due to formScrollSpy makeing every section the active one wherer the div > window ( you get the section by the each function
and the deach funciton is making the current section active for the second and third seciton.  It should only make it active for this section )t