NCIATWP-1015 Enhancements from the June 12, 2017 Meeting
Enlarge print icon and place one at the bottom as well as
top of the page. Include this on every breakpoint.

Dev team collaboration – work on a print style sheet.  Created a print
style sheet, but need to test it for the Ineternet Explorer on the sandbox
and/or dev ( Ask Brian Park).