NCIATWP-1046	Change the width of the Result Boxes
Usability

Dev team collaboration – consider making width of purple boxes on Results pages smaller since there can be a lot of purple background at some breakpoints.

For the purple boxes on the results pages, could we please make the width 85% at the 1025 breakpoint and higher? Margin should be auto so they are still centered on the page. Leave the width at 100% for 1024 and smaller.

Developer
Will need to adjust the widths at the correct breakpoint and center it.

Nervous, about centering it is always a little tricky