NCIATWP-1055		Adjust the Buttons and Links on the Results Page
Usability Experts

“Edit Responses” and “Start New Assessment”

Michelle to consider placement and color of buttons first.

Then CBIIT to add/update both buttons on bottom of all Results pages

Remove “Edit Responses” from top so users can consume info and then change responses if necessary using bottom buttons (CBIIT)
•	Remove Edit Response
•	Put the Printer Icon so the Same Line as Personalized Results.
•	Change the Color of Edit Responses to Purple and put it on its own line. Adjust the size of the button and the Font
•	Put it on its own line centered and Change the font and the color.
•	Put the PDF and Print Icon (centered) on their own line

Enlarge print icon and place one at the bottom as well as top of the page. Include this on every breakpoint.
This was not part of the spec, so I followed the spec on this one.