NCIATWP-1078 BCRAT: Warning pop-up disclaimers
Usability Engineer

BCRAT: Warning pop-up disclaimers that allow you to continue with the
tool, but indicate this this may not be right for you

Remove pop-ups from the question “What is the patient’s race/ethnicity?”
for Hispana/Latina and American Indian or Alaskan Native (CBIIT)

As discussed in 6/12/2018 meeting, this text should be included on the
homepage and the results page. Mitch/Jennifer to add a line to
the homepage text about other races

Developer

Need information about the text that I need be in the landing page
and the results page. I have not been able to find it yet.

Comment
They said that the disclaimers would be one the Result Page. After
looking at the spec, I found that on results page under the
"Your Answers" Section for following question "  What is the patient’s
race/ethnicity?" the disclaimer was under it.

The disclaimer is only needed from some of the answers ( ex. Unknown).