NCIATWP-988    Web Analytics - Add Adobe DAP code to RAT Tools
Instructions
Adding DTM to the host site requires the addition of two script tags to the source code of the site. These tags must be placed on every page of the site. In many cases, pages are generated dynamically and the script may only need to go in a few places within the code to ensure this requirement is met, although the site administrator should determine which files should be modified based on your specific environment.
ODDC is happy to provide any assistance needed, including help with testing the code before release, and verifying that DAP is reporting correctly to GSA after the implementation. If you have any questions, or would like assistance, please contact Ilene France.
Javascript Tags
The following tags must be added to all pages of the website:
1. The first line of the <head> tag should be the following script tag:
<script src="//assets.adobedtm.com/f1bfa9f7170c81b1a9a9ecdcc6c5215ee0b03c84/satelliteLib-4b219b82c4737db0e1797b6c511cf10c802c95cb.js"></script>
2. The last line of the body, just prior to the closing </body> tag, should be the following script tag:
<script type="text/javascript">_satellite.pageBottom();</script>

Reply